Initial commit for global repo
This repository is intended to become the global home for Babelfish on
Github. We need such a repo because we're forced into having 2 separate
repos (one for the modified engine code, and one for he actual babelfish
extension code). It will also contain build tools and some other global
items.

Signed-off-by: Jim C. Nasby <nasbyj@amazon.com>